+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Added
+- Added the `php craft tinymce/config/regenerate-default` console command
+- Added `spicyweb\tinymce\console\controllers\ConfigController`
+- Added `spicyweb\tinymce\services\ConfigService`
+
 ## 1.0.1 - 2022-09-26
 
 ### Fixed